Update SimpleUtils.java
Because blocks.getEntriesView is backed by a set and not a list, so I think toList is giving an error. This should be able to be resolved by using the Collector API instead as was previously standard.

This also might only be an issue with loading the roses mod, so it might be better to make a new function if this function is in use and working elsewhere in other repos.